+++ hosts/tau/home.nix
@@ -15,16 +15,15 @@ in {
     "Xcursor.size" = 48;
   };
 
-  xsession.windowManager.i3.extraConfig = ''
-    workspace 101 output eDP1
+  xsession.windowManager.i3.config.startup = [
+    { command = "xrandr --output eDP1 --scale '1.6x1.6'"; notification = false; }
+    { command = "nm-applet"; notification = false; }
+    { command = "slack"; notification = false; }
+  ];
 
-    exec --no-startup-id i3 workspace 101
-
-    exec --no-startup-id xrandr --output eDP1 --scale '1.6x1.6'
-    exec --no-startup-id nm-applet
-    exec --no-startup-id xss-lock -- bash -c "i3lock -c 000000; xset dpms force off"
-    exec --no-startup-id slack
-  '';
+  xsession.windowManager.i3.config.workspaceOutputAssign = [
+    { workspace = "101"; output = "eDP1"; }
+  ];
 
   services.polybar = {
     config = {