+++ modules/samba.nix
@@ -1,6 +1,7 @@
 {
-  nixosModule = { config, ... }: {
+  nixosModule = { ... }: {
     services.samba.enable = true;
+    services.samba.openFirewall = true;
     services.samba.shares.phi = {
       path = "/";
       "read only" = "no";
@@ -10,8 +11,5 @@
       "force user" = "enzime";
       "force group" = "users";
     };
-
-    networking.firewall.allowedTCPPorts = (assert (!builtins.hasAttr "openFirewall" config.services.samba); [ 139 445 ]);
-    networking.firewall.allowedUDPPorts = (assert (!builtins.hasAttr "openFirewall" config.services.samba); [ 137 138 ]);
   };
 }