+++ overlays/vscode-extensions.nix
@@ -1,12 +1,12 @@
 self: super:
 
 let
-  inherit (super.lib) hasAttrByPath recursiveUpdate;
+  inherit (super.lib) getVersion hasAttrByPath recursiveUpdate versionOlder;
   inherit (super.vscode-utils) buildVscodeMarketplaceExtension;
 in {
   vscode-extensions = recursiveUpdate super.vscode-extensions {
-    # The `assert` ensures that the extension isn't already present in `nixpkgs`
-    asvetliakov.vscode-neovim = (assert (!hasAttrByPath ["asvetliakov" "vscode-neovim"] super.vscode-extensions); buildVscodeMarketplaceExtension {
+    # Only override if version in `nixpkgs` is older
+    asvetliakov.vscode-neovim = (assert versionOlder (getVersion super.vscode-extensions.asvetliakov.vscode-neovim) "0.0.83"; buildVscodeMarketplaceExtension {
       mktplcRef = {
         name = "vscode-neovim";
         publisher = "asvetliakov";
@@ -15,6 +15,7 @@ in {
       };
     });
 
+    # The `assert` ensures that the extension isn't already present in `nixpkgs`
     ethansk.restore-terminals = (assert (!hasAttrByPath ["ethansk" "restore-terminals"] super.vscode-extensions); buildVscodeMarketplaceExtension {
       mktplcRef = {
         name = "restore-terminals";
@@ -24,30 +25,12 @@ in {
       };
     });
 
-    kamikillerto.vscode-colorize = (assert (!hasAttrByPath ["kamikillerto" "vscode-colorize"] super.vscode-extensions); buildVscodeMarketplaceExtension {
-      mktplcRef = {
-        name = "vscode-colorize";
-        publisher = "kamikillerto";
-        version = "0.11.1";
-        sha256 = "1h82b1jz86k2qznprng5066afinkrd7j3738a56idqr3vvvqnbsm";
-      };
-    });
-
     rioj7.commandOnAllFiles = (assert (!hasAttrByPath ["rioj7" "commandOnAllFiles"] super.vscode-extensions); buildVscodeMarketplaceExtension {
       mktplcRef = {
         name = "commandOnAllFiles";
         publisher = "rioj7";
         version = "0.3.0";
         sha256 = "04f1sb5rxjwkmidpymhqanv8wvp04pnw66098836dns906p4gldl";
-      };
-    });
-
-    shardulm94.trailing-spaces = (assert (!hasAttrByPath ["shardulm94" "trailing-spaces"] super.vscode-extensions); buildVscodeMarketplaceExtension {
-      mktplcRef = {
-        publisher = "shardulm94";
-        name = "trailing-spaces";
-        version = "0.3.1";
-        sha256 = "0h30zmg5rq7cv7kjdr5yzqkkc1bs20d72yz9rjqag32gwf46s8b8";
       };
     });
   };